Functioning API calls and iteration
Any mishandling of the header during the call to the petfinder api
results in an invalidated token which needs regenerated. This has been
the source of almost every problem to date getting up and running. Now
calls are successful and the structure is in place to generate divs to
hold the returned data.