Breed list sources separated to dropdownArrays
Lists for breeds has been moved to a separate file for separation of
concerns and readability. This broke functionality until the script
element importing the file added the tag 'type="module"'. This broke the
 scroll to top button, so its function call was removed from HTML and an
event listener has been added instead. This clears way for adding a pet
type toggle button and displaying different types of animals.