Add regex function to validate input field
While MongoDB has its own protections, and I have previously disabled
put, post, patch, and delete methods, the input field on the front end
now only accepts alphanumeric input. If anything else is present at the
time of search, it is ignored and the rest of the results are displayed
along with an alert function informing the user.