Search now stops with invalid input
Instead of merely ignoring the search field, the search is aborted all-
together. This seems to be a better solution than ignoring what the user
 is attempting to accomplish and doing something else. Event listeners
have been moved to the end of the search.js code. Logic has been changed
 on the event listener as well as applySelectedCriteria. The colors of
the return to top button have been changed as well to show up better
against the footer colors.