Revert repo to release version
Workflows deleted, webpack gone, test code to trick build computers has
been deleted, and the repository has returned to the state of
functionality in the state of it's turn-in form. Front end code was
duplicated and moved to another repository that can behave. README was
changed to reflect the new deployment URL.